Fixes panelRenameTest and move milestone and panelrename test to stable
