+++ src/test/java/guitests/MilestonePickerTests.java
@@ -1,6 +1,5 @@
-package unstable;
+package guitests;
 
-import guitests.UITest;
 import javafx.scene.input.KeyCode;
 import org.junit.Test;
